refactor(SpokePoolClient): Relocate validateFillForDeposit
This function compares a Deposit with a Fill and doesn't relate to the
SpokePool or SpokePoolClient. Move it to FillUtils instead.
